+++ app/src/main/java/ru/practicum/android/diploma/main/data/network/NetworkClientImpl.kt
@@ -3,7 +3,6 @@ package ru.practicum.android.diploma.main.data.network
 import android.content.Context
 import android.net.ConnectivityManager
 import android.net.NetworkCapabilities
-import android.util.Log
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import retrofit2.HttpException
@@ -31,7 +30,6 @@ class NetworkClientImpl(
     private suspend fun apiCall(
         doRequest: suspend () -> Any?
     ): NetworkResponse {
-        Log.d("NetworkClient", "token = $token")
         if (!isConnected()) {
             return NetworkResponse().apply { resultCode = NetworkResponse.NO_CONNECTION }
         }
@@ -43,7 +41,6 @@ class NetworkClientImpl(
                     data = result
                 }
             } catch (ex: HttpException) {
-                Log.e("error", "Error: ${ex.message}")
                 NetworkResponse().apply { resultCode = NetworkResponse.BAD_REQUEST }
             }
         }
