+++ app/src/main/java/ru/practicum/android/diploma/main/ui/SearchScreen.kt
@@ -4,25 +4,17 @@ package ru.practicum.android.diploma.main.ui
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentSize
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -39,7 +31,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.painterResource
@@ -49,12 +40,8 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import coil3.ImageLoader
-import coil3.compose.AsyncImage
 import org.koin.androidx.compose.koinViewModel
-import org.koin.compose.koinInject
 import ru.practicum.android.diploma.R
-import ru.practicum.android.diploma.main.domain.models.VacancyCard
 import ru.practicum.android.diploma.ui.navigation.ActionBack
 import ru.practicum.android.diploma.ui.navigation.ActionFilter
 import ru.practicum.android.diploma.ui.navigation.AppBarTop
@@ -154,16 +141,7 @@ fun SearchScreen(
 private fun SearchContent(state: SearchState, viewModel: SearchViewModel) {
     when {
         state.isLoading -> {
-            Box(
-                modifier = Modifier
-                    .fillMaxSize(),
-                contentAlignment = Alignment.Center
-            ) {
-                CircularProgressIndicator(
-                    color = MaterialTheme.colorScheme.primary,
-                    modifier = Modifier.size(48.dp)
-                )
-            }
+            LoadingIndicator()
         }
 
         state.isNetworkError -> {
@@ -237,106 +215,11 @@ private fun SearchContent(state: SearchState, viewModel: SearchViewModel) {
         }
 
         state.vacancies.isNotEmpty() -> {
-            Column(
-                modifier = Modifier.fillMaxSize(),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Surface(
-                    color = MaterialTheme.colorScheme.primary,
-                    shape = RoundedCornerShape(12.dp),
-                    modifier = Modifier
-                        .wrapContentSize()
-                        .padding(top = 4.dp, bottom = 8.dp),
-                ) {
-                    Text(
-                        text = "Найдено ${state.totalCount} вакансий",
-                        color = MaterialTheme.colorScheme.onPrimary,
-                        style = MaterialTheme.typography.bodyMedium,
-                        modifier = Modifier.padding(horizontal = 12.dp, vertical = 4.dp)
-                    )
-                }
-
-                LazyColumn {
-                    items(state.vacancies, key = { it.id }) { vacancy ->
-                        VacancyCard(
-                            vacancy = vacancy,
-                            onClick = { viewModel.onVacancyClick(vacancy.id) }
-                        )
-                        if (state.vacancies.lastOrNull() == vacancy && state.hasMorePages && !state.isLoading) {
-                            LaunchedEffect(Unit) {
-                                viewModel.loadNextPage()
-                            }
-                        }
-                    }
-                }
-            }
+            VacanciesSearchResult(state, viewModel)
         }
 
         state.searchText.isEmpty() -> {
-            Box(
-                modifier = Modifier
-                    .wrapContentSize()
-                    .padding(horizontal = 16.dp),
-                contentAlignment = Alignment.Center
-            ) {
-                Image(
-                    modifier = Modifier.fillMaxSize(),
-                    alignment = Alignment.Center,
-                    painter = painterResource(R.drawable.job_search),
-                    contentDescription = "Ищу работу"
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun VacancyCard(
-    vacancy: VacancyCard,
-    onClick: () -> Unit,
-    imageLoader: ImageLoader = koinInject()
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable { onClick() }
-            .padding(horizontal = 16.dp, vertical = 8.dp),
-    ) {
-        AsyncImage(
-            model = vacancy.logo,
-            contentDescription = "Логотип компании",
-            imageLoader = imageLoader,
-            modifier = Modifier
-                .size(48.dp)
-                .clip(RoundedCornerShape(12.dp)),
-            placeholder = painterResource(R.drawable.ic_placeholder_32),
-            error = painterResource(R.drawable.ic_placeholder_32)
-        )
-
-        Spacer(modifier = Modifier.width(12.dp))
-
-        Column(modifier = Modifier.weight(1f)) {
-            Text(
-                text = buildString {
-                    append(vacancy.name)
-                    vacancy.city?.let { append(", $it") }
-                },
-                style = MaterialTheme.typography.titleMedium,
-                maxLines = 3
-            )
-
-            Text(
-                text = vacancy.company ?: "",
-                style = MaterialTheme.typography.bodyLarge,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                maxLines = 1
-            )
-            Text(
-                text = vacancy.salary ?: stringResource(R.string.salary_not_specified),
-                style = MaterialTheme.typography.bodyLarge,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                maxLines = 1
-            )
+            EmptySearchPlaceholder()
         }
     }
 }
