+++ app/src/main/java/ru/practicum/android/diploma/main/ui/VacancyCard.kt
@@ -1,0 +1,76 @@
+@file:Suppress("MagicNumber")
+
+package ru.practicum.android.diploma.main.ui
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import coil3.ImageLoader
+import coil3.compose.AsyncImage
+import org.koin.compose.koinInject
+import ru.practicum.android.diploma.R
+import ru.practicum.android.diploma.main.domain.models.VacancyCard
+
+@Composable
+fun VacancyCard(
+    vacancy: VacancyCard,
+    onClick: () -> Unit,
+    imageLoader: ImageLoader = koinInject()
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .clickable { onClick() }
+            .padding(horizontal = 16.dp, vertical = 8.dp),
+    ) {
+        AsyncImage(
+            model = vacancy.logo,
+            contentDescription = "Логотип компании",
+            imageLoader = imageLoader,
+            modifier = Modifier
+                .size(48.dp)
+                .clip(RoundedCornerShape(12.dp)),
+            placeholder = painterResource(R.drawable.ic_placeholder_32),
+            error = painterResource(R.drawable.ic_placeholder_32)
+        )
+
+        Spacer(modifier = Modifier.width(12.dp))
+
+        Column(modifier = Modifier.weight(1f)) {
+            Text(
+                text = buildString {
+                    append(vacancy.name)
+                    vacancy.city?.let { append(", $it") }
+                },
+                style = MaterialTheme.typography.titleMedium,
+                maxLines = 3
+            )
+            Text(
+                text = vacancy.company ?: "",
+                style = MaterialTheme.typography.bodyLarge,
+                color = MaterialTheme.colorScheme.onSurfaceVariant,
+                maxLines = 1
+            )
+            Text(
+                text = vacancy.salary ?: stringResource(R.string.salary_not_specified),
+                style = MaterialTheme.typography.bodyLarge,
+                color = MaterialTheme.colorScheme.onSurfaceVariant,
+                maxLines = 1
+            )
+        }
+    }
+}
